requirements.txt
beautifulsoup4
chromadb
cohere
datasets
deepeval
langchain
langchain-community
langchain-openai
langchainhub
numpy
pytube
python-dotenv
ragas
ragatouille
requests
tiktoken
youtube-transcript-api
git+https://github.com/j-min/GROUSe#egg=grouse-eval